requirements.txt
# ENV set up
python-dotenv

# util
pyfunctional

# torch
torch
torchvision

# hugging face stuff
transformers

# text tokenizer
sentencepiece

# qunatizer
accelerate
bitsandbytes

# Document text extraction
python-doctr

#embedding
sentence-transformers